# Makes IonRawBinaryWriter's currentFieldSid an int primitive instead of an Integer