…cutting down on allocations and garbage caused by autoboxing.

*Description of changes:*

`currentFieldSid` gets set every time the user sets a field name on the binary writer, which must happen before every value that is contained in a struct. Previously, an `Integer` was used to hold the current field symbol ID, with `null` indicating that no symbol ID had been assigned to the current value. However, because field name symbol IDs must be positive, a negative number can be used for this purpose, allowing for an `int` primitive to be used.

Boxing an `int` into an `Integer` each time a field name is set can have a real impact on applications that produce long streams of binary Ion data. Profiling such an application revealed that `Integer` boxing from this code path was IonJava's largest source of allocations, and was a top 5 allocation contributor for the application as a whole.

In a benchmark that wrote a stream of structs that each contained many field names, this change resulted in a modest speedup (12.6%) and reduction in heap usage (17%), along with a complete elimination of garbage collections during the trials. It should be noted that this benchmark's structs contain many more field names and much less value diversity (only Ion integers) than the typical real-world Ion struct. As such, these results should be analyzed for their impact on this particular aspect of writing binary Ion streams; the speedups and GC churn reductions will not be this high for a typical Ion stream.

```
Old:

Benchmark                                                                    Mode  Cnt          Score          Error   Units
IonJavaContainersBenchmark.fieldNameSids                                     avgt   20       1276.990 ±       69.207   ms/op
IonJavaContainersBenchmark.fieldNameSids:Heap usage                          avgt   20        949.725 ±      172.895      MB
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Eden Space}           avgt   20        623.813 ±       86.107  MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Eden Space}.norm      avgt   20  654114917.807 ± 90289884.879    B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Old Gen}              avgt   20         27.146 ±       49.914  MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Old Gen}.norm         avgt   20   28465160.169 ± 52338859.328    B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Survivor Space}       avgt   20         11.594 ±       31.118  MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Survivor Space}.norm  avgt   20   12157463.205 ± 32630029.256    B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.count                           avgt   20         29.000                 counts
IonJavaContainersBenchmark.fieldNameSids:·gc.time                            avgt   20       1229.000                     ms

New:

Benchmark                                                                    Mode  Cnt     Score    Error   Units
IonJavaContainersBenchmark.fieldNameSids                                     avgt   20  1115.598 ± 61.767   ms/op
IonJavaContainersBenchmark.fieldNameSids:Heap usage                          avgt   20   784.547 ±  0.013      MB
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Eden Space}           avgt   20       ≈ 0           MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Eden Space}.norm      avgt   20       ≈ 0             B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Old Gen}              avgt   20       ≈ 0           MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Old Gen}.norm         avgt   20       ≈ 0             B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Survivor Space}       avgt   20       ≈ 0           MB/sec
IonJavaContainersBenchmark.fieldNameSids:·gc.churn.{PS Survivor Space}.norm  avgt   20       ≈ 0             B/op
IonJavaContainersBenchmark.fieldNameSids:·gc.count                           avgt   20       ≈ 0           counts
IonJavaContainersBenchmark.fieldNameSids:·gc.time                            avgt   20       ≈ 0               ms
``` 

These results were generated by the following JMH benchmark code:

```java
import com.amazon.ion.IonType;
import com.amazon.ion.IonWriter;
import com.amazon.ion.impl.bin._Private_IonManagedWriter;
import com.amazon.ion.impl.bin._Private_IonRawWriter;
import com.amazon.ion.system.IonBinaryWriterBuilder;
import org.openjdk.jmh.annotations.Benchmark;
import org.openjdk.jmh.annotations.Level;
import org.openjdk.jmh.annotations.Mode;
import org.openjdk.jmh.annotations.Scope;
import org.openjdk.jmh.annotations.Setup;
import org.openjdk.jmh.annotations.State;
import org.openjdk.jmh.annotations.TearDown;
import org.openjdk.jmh.infra.BenchmarkParams;
import org.openjdk.jmh.infra.IterationParams;
import org.openjdk.jmh.profile.GCProfiler;
import org.openjdk.jmh.profile.InternalProfiler;
import org.openjdk.jmh.profile.ProfilerResult;
import org.openjdk.jmh.results.AggregationPolicy;
import org.openjdk.jmh.results.IterationResult;
import org.openjdk.jmh.results.Result;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.options.OptionsBuilder;

import java.io.IOException;
import java.io.OutputStream;
import java.lang.management.ManagementFactory;
import java.util.ArrayList;
import java.util.Collection;
import java.util.List;
import java.util.concurrent.TimeUnit;

@State(Scope.Benchmark)
public class IonJavaFieldNameBenchmark {

    private IonWriter writer;
    private _Private_IonRawWriter rawWriter;
    private int bytesWritten = 0;

    @Setup(Level.Iteration)
    public void setup() {
        writer = IonBinaryWriterBuilder.standard().build(new OutputStream() {
            @Override
            public void write(int b) {
                bytesWritten++;
            }
        });
        rawWriter = writer.asFacet(_Private_IonManagedWriter.class).getRawWriter();
    }

    @TearDown(Level.Iteration)
    public void tearDown() throws IOException {
        writer.close();
        bytesWritten = 0;
    }

    @Benchmark
    public int fieldNameSids() throws Exception {
        for (int j = 0; j < 5; j++) {
            rawWriter.stepIn(IonType.STRUCT);
            for (int i = 0; i < 10000000; i++) {
                rawWriter.setFieldNameSymbol(i);
                rawWriter.writeInt(i);
            }
            rawWriter.stepOut();
        }
        writer.finish();
        return bytesWritten;
    }

    public static class HeapProfiler implements InternalProfiler {

        @Override
        public boolean checkSupport(List<String> list) {
            return true;
        }

        @Override
        public String label() {
            return "HEAP_USAGE";
        }

        @Override
        public String getDescription() {
            return "Heap usage profiler";
        }

        @Override
        public void beforeIteration(BenchmarkParams benchmarkParams, IterationParams iterationParams) {
            // Nothing.
        }

        @Override
        public Collection<? extends Result> afterIteration(BenchmarkParams benchmarkParams, IterationParams iterationParams,
                                                           IterationResult result) {

            long totalHeap = ManagementFactory.getMemoryMXBean().getHeapMemoryUsage().getUsed();

            Collection<Result> results = new ArrayList<>();
            results.add(new ProfilerResult("Heap usage", totalHeap / 1e6, "MB", AggregationPolicy.AVG));

            return results;
        }
    }

    public static void main(String[] args) throws Exception {
        new Runner(
            new OptionsBuilder()
                .include(IonJavaFieldNameBenchmark.class.getSimpleName())
                .addProfiler(HeapProfiler.class)
                .addProfiler(GCProfiler.class)
                .mode(Mode.AverageTime)
                .measurementIterations(10)
                .warmupIterations(10)
                .forks(2)
                .shouldDoGC(true)
                .timeUnit(TimeUnit.MILLISECONDS)
                .build()
        ).run();
    }
}
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.